CHANGE: renamed single contracts sub folder for clarity
