+++ src/controllers/api/index.ts
@@ -7,8 +7,8 @@ export * from "./ContractSubControllers";
 export * from "./ContractsController";
 export * from "./PenalitiesController";
 export * from "./PortfolioSubControllers";
-export * from "./contracts/ContractTransactionsCreditController";
-export * from "./contracts/ContractTransactionsDebitController";
-export * from "./contracts/ContractViolationsController";
+export * from "./contract/ContractTransactionsCreditController";
+export * from "./contract/ContractTransactionsDebitController";
+export * from "./contract/ContractViolationsController";
 export * from "./portfolio/PortfolioTransactionsCreditController";
 export * from "./portfolio/PortfolioTransactionsDebitController";
